Split the Sampler into procfs and cgroup focus
This commit splits the Sampler so that cgroup collection is in a separate
implementation from procfs, which has gotten sprawling. I have maintained the
existing Sampler interface and hidden the two new implementations inside of it,
although we might choose to expose them at some point in the future.
